+++ src/main/java/tw/com/cathaybk/repository/EmployeeDAO.java
@@ -1,4 +1,4 @@
-package tw.com.cathaybk.dao;
+package tw.com.cathaybk.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
